+++ packages/echo-core/CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## v0.6.19
+
+- Updated to use latest echo-base 0.6.12.
+- Updated nuget packages
+
 ## v0.6.18
 
 - Updated to use latest echo-base 0.6.11.